+++ doc/changes.rst
@@ -2,7 +2,7 @@
 prospect's Change Log
 =====================
 
-1.0.2 (unreleased)
+1.0.2 (2021-01-27)
 ------------------
 
 * Optional countdown widget (PR `#58`_).